# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "AlertActionCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "AlertActionCreate",
        "type": "object"
      },
...
      "AlertDetailResponse": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/AlertActionResponse"
            },
            "title": "Actions",
            "type": "array"
          },
          "alert_id": {
            "format": "uuid",
            "title": "Alert Id",
            "type": "string"
          },
          "alert_type": {
            "title": "Alert Type",
            "type": "string"
          },
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "first_triggered_at": {
            "format": "date-time",
            "title": "First Triggered At",
            "type": "string"
          },
          "last_triggered_at": {
            "format": "date-time",
            "title": "Last Triggered At",
            "type": "string"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "alert_id",
          "campaign_id",
          "alert_type",
          "severity",
          "status",
          "first_triggered_at",
          "last_triggered_at",
          "occurrence_count",
          "version",
          "actions"
        ],
        "title": "AlertDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TimelineEventResponse": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "event_time": {
            "format": "date-time",
            "title": "Event Time",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "event_time",
          "data"
        ],
        "title": "TimelineEventResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Event-time-aware pump-and-dump intelligence backend.",
    "title": "Scam2Market Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts/{alert_id}": {
      "get": {
        "operationId": "alert_detail_api_v1_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alert Detail",
        "tags": [
          "analyst"
        ]
      }
    },
    "/api/v1/alerts/{alert_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_alert_api_v1_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertActionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge Alert",
        "tags": [
          "analyst"
        ]
      }
    },
...
    "/api/v1/assets/{asset_id}/narratives": {
      "get": {
        "operationId": "asset_narratives_api_v1_assets__asset_id__narratives_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "default": "LIVE",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NarrativeResponse"
                  },
                  "title": "Response Asset Narratives Api V1 Assets  Asset Id  Narratives Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Narratives",
        "tags": [
          "analyst"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/overview": {
      "get": {
        "operationId": "asset_overview_api_v1_assets__asset_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "default": "LIVE",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Overview",
        "tags": [
          "analyst"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/scores": {
      "get": {
        "operationId": "asset_scores_api_v1_assets__asset_id__scores_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScoreResponse"
                  },
                  "title": "Response Asset Scores Api V1 Assets  Asset Id  Scores Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Scores",
        "tags": [
          "analyst"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/timeline": {
      "get": {
        "operationId": "asset_timeline_api_v1_assets__asset_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "default": "LIVE",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimelineEventResponse"
                  },
                  "title": "Response Asset Timeline Api V1 Assets  Asset Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Timeline",
        "tags": [
          "analyst"
        ]
      }
    },
...
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "operationId": "campaign_detail_api_v1_campaigns__campaign_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Campaign Detail",
        "tags": [
          "analyst"
        ]
      }
    },
...
    "/api/v1/campaigns/{campaign_id}/graph": {
      "get": {
        "operationId": "campaign_graph_api_v1_campaigns__campaign_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Campaign Graph",
        "tags": [
          "analyst"
        ]
      }
    },
...
    "/api/v1/narratives/{narrative_id}": {
      "get": {
        "operationId": "narrative_detail_api_v1_narratives__narrative_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "narrative_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Narrative Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Narrative Detail",
        "tags": [
          "analyst"
        ]
      }
    },
...
    "/api/v1/watchlists": {
      "get": {
        "operationId": "list_watchlists_api_v1_watchlists_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "default": "LIVE",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistResponse"
                  },
                  "title": "Response List Watchlists Api V1 Watchlists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Watchlists",
        "tags": [
          "analyst"
        ]
      },
      "post": {
        "operationId": "create_watchlist_api_v1_watchlists_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watchlist",
        "tags": [
          "analyst"
        ]
      }
    },
...
  }
}